+++ ProjectArtifacts/CommonAssemblyInfo.cs
@@ -2,8 +2,8 @@ using System.Reflection;
 
 [assembly: AssemblyCompany("Sean O'Donnell")]
 [assembly: AssemblyProduct("TkMemory")]
-[assembly: AssemblyCopyright("Copyright © 2019")]
+[assembly: AssemblyCopyright("Copyright © 2020")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
-[assembly: AssemblyVersion("1.1.0")]
+[assembly: AssemblyVersion("1.1.1")]