+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2020-04-17
+### Added
+- NuGet icon to Project Artifacts
+### Changed
+- Movement command cooldown is now 167 milliseconds to alleviate AHK memory errors from too many commands
+### Fixed
+- `Haiku of Quantao` corrected to `Haiku of Qantao` in Poet key spell priorities
+- Status checks for ASV now properly check for `Rhyme of Walsuk`, `Sonnet of ReShor`, and `Diancecht's Follain`
+
 ## [1.1.0] - 2019-12-08
 ### Added
 - Added commands for moving the player and for following another multibox player
@@ -13,5 +22,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - :tada: Initial commit!
 
-[1.1.0]: https://github.com/elodon/TkMemory/commit/master
+[1.1.1]: https://github.com/elodon/TkMemory/commit/master
+[1.1.0]: https://github.com/elodon/TkMemory/commit/6166f58591d5a3ed24f2db30ef1bec1c1e84d93a
 [1.0.0]: https://github.com/elodon/TkMemory/commit/e4043a44b6a79430a1307f0ef8b0a5ddaf068999